refactor: move connector to internal/connector
